restructure layers directory
this change introduces sub-directories to provide a single module for a group of layers rather than dumping them in a single file with the same as the sub-directory
