+++ pytorch_forecasting/layers/__init__.py
@@ -2,11 +2,7 @@
 Architectural deep learning layers from `nn.Module`.
 """
 
-from pytorch_forecasting.layers.attention import (
-    AttentionLayer,
-    FullAttention,
-    TriangularCausalMask,
-)
+from pytorch_forecasting.layers.attention import AttentionLayer, FullAttention
 from pytorch_forecasting.layers.embeddings import (
     DataEmbedding_inverted,
     EnEmbedding,
@@ -16,7 +12,7 @@ from pytorch_forecasting.layers.encoders import (
     Encoder,
     EncoderLayer,
 )
-from pytorch_forecasting.layers.output_layers import (
+from pytorch_forecasting.layers.output.flatten_head import (
     FlattenHead,
 )
