+++ pytorch_forecasting/layers/attention/__init__.py
@@ -1,0 +1,8 @@
+"""
+Attention Layers for pytorch-forecasting models.
+"""
+
+from pytorch_forecasting.layers.attention.attention_layer import AttentionLayer
+from pytorch_forecasting.layers.attention.full_attention import FullAttention
+
+__all__ = ["AttentionLayer", "FullAttention"]
