+++ pytorch_forecasting/layers/attention/attention_layer.py
@@ -10,67 +10,6 @@ import torch.nn as nn
 import torch.nn.functional as F
 
 
-class TriangularCausalMask:
-    """
-    Triangular causal mask for attention mechanism.
-    """
-
-    def __init__(self, B, L, device="cpu"):
-        mask_shape = [B, 1, L, L]
-        with torch.no_grad():
-            self._mask = torch.triu(
-                torch.ones(mask_shape, dtype=torch.bool), diagonal=1
-            ).to(device)
-
-    @property
-    def mask(self):
-        return self._mask
-
-
-class FullAttention(nn.Module):
-    """
-    Full attention mechanism with optional masking and dropout.
-    Args:
-        mask_flag (bool): Whether to apply masking.
-        factor (int): Factor for scaling the attention scores.
-        scale (float): Scaling factor for attention scores.
-        attention_dropout (float): Dropout rate for attention scores.
-        output_attention (bool): Whether to output attention weights."""
-
-    def __init__(
-        self,
-        mask_flag=True,
-        factor=5,
-        scale=None,
-        attention_dropout=0.1,
-        output_attention=False,
-    ):
-        super(FullAttention, self).__init__()
-        self.scale = scale
-        self.mask_flag = mask_flag
-        self.output_attention = output_attention
-        self.dropout = nn.Dropout(attention_dropout)
-
-    def forward(self, queries, keys, values, attn_mask, tau=None, delta=None):
-        B, L, H, E = queries.shape
-        _, S, _, D = values.shape
-        scale = self.scale or 1.0 / sqrt(E)
-
-        scores = torch.einsum("blhe,bshe->bhls", queries, keys)
-
-        if self.mask_flag:
-            if attn_mask is None:
-                attn_mask = TriangularCausalMask(B, L, device=queries.device)
-            scores.masked_fill_(attn_mask.mask, -np.abs)
-        A = self.dropout(torch.softmax(scale * scores, dim=-1))
-        V = torch.einsum("bhls,bshd->blhd", A, values)
-
-        if self.output_attention:
-            return V.contiguous(), A
-        else:
-            return V.contiguous(), None
-
-
 class AttentionLayer(nn.Module):
     """
     Attention layer that combines query, key, and value projections with an attention
