+++ pytorch_forecasting/layers/output/__init__.py
@@ -1,0 +1,7 @@
+"""
+Implementation of output layers for PyTorch Forecasting.
+"""
+
+from .flatten_head import FlattenHead
+
+__all__ = ["FlattenHead"]
